iiod: Fix invalid print format
time_t is implementation defined, and therefore there is no printf
format specifier that is guaranteed to work on every platform.

Address this issue by using %ld and casting the value to long.

+++ iiod/dns-sd.c
@@ -320,7 +320,8 @@ static void start_avahi_thd(struct thread_pool *pool, void *d)
 		if (avahi.client)
 			break;
 again:
-		IIO_INFO("Avahi didn't start, try again in %d seconds later\n", ts.tv_sec);
+		IIO_INFO("Avahi didn't start, try again in %ld seconds later\n",
+			 (long)ts.tv_sec);
 		nanosleep(&ts, NULL);
 		ts.tv_sec++;
 		/* If it hasn't started in 20 times over 210 seconds (3.5 min),
